refactor(solver): decompose DefinitionStore god-file along concern axes
Split the 2291-line def/core.rs (largest file in the repo) into composed
sub-stores and concern-scoped child modules, all behavior-preserving:

- DefStateFlags (def/core/state_flags.rs): the six poison / circular /
  publish-isolation / alias-body DashSets, with their invalidation contract
  stated in one place instead of fused as loose fields.
- CrossFileQueryCache (def/core/cross_file_cache.rs): the cross-file query
  memo, its scope stamp, and per-file delegation locks.
- observability (def/core/observability.rs): StoreStatistics value type plus
  the statistics()/estimated_size_bytes() snapshot methods (telemetry only).
- semantic_construction (def/core/semantic_construction.rs): the
  from_semantic_defs* binder-identity -> DefId factories.

DefinitionStore holds the two sub-stores as fields and delegates; the public
method surface is unchanged. def_tests.rs now imports PropertyInfo directly
instead of via the parent module's incidental re-export.

def/core.rs 2291 -> 1675 lines (under the 2000-line cap).

Refs #13417
